Numba Ahead-Of-Time
Numba code is built when the module is called for the first time.
This way the code will be optimized for the processor at hand
instead of more generic prebuild package. In following import
calls build step is not needed.

This means that numba is still a requirement.
